# lcm-engine/lcm_engine/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: |
    The Life-Cycle Management Engine API specification for the Gaia-X project.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: LCM Engine API
  version: 2.3.0
servers:
- description: Testing LCM Engine server 1
  url: https://demo.gxfs.xlab.si
- description: Testing LCM Engine server 2
  url: https://orchestrator.gxfs.xlab.si
- description: Local development
  url: http://localhost:8080
...
paths:
  /auth/logout:
    post:
      operationId: auth_logout
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationStatus'
          description: Authentication status -- always unauthenticated
      summary: Log out of the application
      tags:
      - user
      x-openapi-router-controller: lcm_engine.controllers.user_controller
  /auth/status:
    get:
      operationId: auth_status
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationStatus'
          description: Authentication status
      summary: Get the authentication status
      tags:
      - user
      x-openapi-router-controller: lcm_engine.controllers.user_controller
...
  /workspace/{workspaceId}/authorizations:
    delete:
      operationId: deauthorize_workspace_user
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/AuthorizationBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
          description: Success
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Deauthorize a user from the workspace
      tags:
      - user
      x-openapi-router-controller: lcm_engine.controllers.user_controller
    get:
      operationId: get_workspace_authorizations
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserIdentifier'
                type: array
          description: A list of users authorized to this workspace
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
      summary: List users authorized to this workspace
      tags:
      - user
      x-openapi-router-controller: lcm_engine.controllers.user_controller
    put:
      operationId: authorize_workspace_user
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/AuthorizationBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
          description: Success
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Authorize a new user to the workspace
      tags:
      - user
      x-openapi-router-controller: lcm_engine.controllers.user_controller
...
components:
...
  requestBodies:
...
    AuthorizationBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkspaceUserAuthorizationRequest'
      description: Authorization specification object
      required: true
...
  schemas:
...
    AuthenticationStatus:
      additionalProperties: false
      description: Authentication status
      example:
        userIdentifier: john.doe@example.com
        isLoggedIn: true
      properties:
        userIdentifier:
          example: john.doe@example.com
          title: userIdentifier
          type: string
        isLoggedIn:
          example: true
          title: isLoggedIn
          type: boolean
      required:
      - isLoggedIn
      - userIdentifier
      title: AuthenticationStatus
      type: object
...
    Error:
      additionalProperties: false
      description: An error
      properties:
        msg:
          title: msg
          type: string
      required:
      - msg
      title: Error
      type: object
    UserIdentifier:
      description: A unique identifier for a user
      example: john.doe@example.com
      title: UserIdentifier
      type: string
    Empty:
      additionalProperties: false
      description: Nothing
      maxItems: 0
      type: object
    WorkspaceUserAuthorizationRequest:
      additionalProperties: false
      description: A request for user de-authorization
      example:
        user: john.doe@example.com
      properties:
        user:
          description: A unique identifier for a user
          example: john.doe@example.com
          title: UserIdentifier
          type: string
      required:
      - user
      title: WorkspaceUserAuthorizationRequest
      type: object
...